refactor(capsule): split security.rs into focused submodules
Keep files under the 1000-line CI threshold and separate concerns.

security.rs (1108 lines) -> security/{mod,manifest_gate,test_gates}.rs:

- security/mod.rs (155): CapsuleSecurityGate trait, IdentityOperation,
  identity_capability_satisfies helper, module wiring.
- security/manifest_gate.rs (787): production ManifestSecurityGate struct +
  impl + its unit tests (~500 lines of test coverage — includes the
  per-invocation home:// and traversal cases added in this PR).
- security/test_gates.rs (183): test-only AllowAllGate + DenyAllGate and
  their coverage tests.

Public surface unchanged: external consumers still `use crate::security::
CapsuleSecurityGate / IdentityOperation / ManifestSecurityGate`, via pub
and pub(crate) re-exports in security/mod.rs. 240 astrid-capsule lib tests
green. No behavior change.

Note on wasm/mod.rs (1750 lines): pre-existing technical debt (1621 on
origin/main, CI exempt via the 'already over 1000' rule). Deferring to a
dedicated reorg PR so this PR's diff stays focused on per-invocation VFS.
